requirements.txt
Flask
flask-cors
python-dotenv
llama-cpp-python
vosk
python-Levenshtein